@@ -1,51 +1,49 @@
 module ContractCmd
   class Resolve < ApplicationCommand
 
-    attr_subobjects :contract, :publisher
-    attr_accessor :counterparty
-    attr_delegate_fields :contract
+    attr_reader :base_contract
 
     validate :resolvable_contract
 
-    def self.find(contract)
-      instance = allocate
-      instance.contract = Contract.find(contract.to_i)
-      instance.publisher = instance.contract.publisher
-      instance.counterparty = instance.contract.counterparty
-      instance
-    end
-
     def initialize(contract)
-      @contract = Contract.find(contract.to_i)
-      @publisher = @contract.publisher
-      @counterparty = @contract.counterparty
+      @base_contract = Contract.find(contract.to_i)
+      add_event(:contract, Event::ContractResolved.new(con_opts(contract)))
+      resolve_escrows
     end
 
-    def transact_before_project
-      contract.status = get_status
-      if contract.status == "lapsed"
-        contract.awarded_to = "publisher"
-        contract.publisher.token_balance += contract.token_value
-      else
-        contract.awarded_to = contract.awardee
-        awardee = contract.send(contract.awardee.to_sym)
-        awardee.token_balance += contract.token_value * 2
+    def user_ids
+      contract&.escrows&.reduce([]) do |acc, esc|
+        acc + esc.users.pluck(:id)
+      end.sort.uniq
+    end
+
+    def resolve_escrows
+      base_contract.escrows.each_with_index do |escrow, idxe|
+        poslst = base_contract.awardee == "fixed" ? escrow.fixed_positions : escrow.unfixed_positions
+        escvol = poslst.map {|p| p.volume}.sum
+        poslst.each_with_index do |position, idxp|
+          prorata = escvol / position.volume.to_f
+          payout  = escrow.total_value * prorata
+          add_event("esc_#{idxe}_#{idxp}", Event::EscrowDistributed.new(base_opts({uuid: escrow.uuid})))
+          add_event("usr_#{idxe}_#{idxp}", Event::UserCredited.new(base_opts({uuid: position.user.uuid, amount: payout})))
+        end
       end
     end
 
     private
 
-    def get_status
-      status = "lapsed"
-      status = "awarded" if contract.counterparty.present?
-      status
+    def con_opts(contract)
+      cmd_opts.merge({uuid: contract.uuid})
+    end
+
+    def base_opts(input)
+      cmd_opts.merge(input)
     end
 
     def resolvable_contract
-      if Time.now < contract.matures_at
-        errors.add(:matures_at, "contract has not matured")
+      if BugmTime.now < base_contract.maturation
+        errors.add(:maturation, "contract has not matured")
       end
     end
   end
 end
-
