@@ -1,21 +1,35 @@
 module UserCmd
   class Create < ApplicationCommand
 
-    attr_subobjects      :user
-    attr_delegate_fields :user
-
-    # delegate :authenticatable_salt, :to => :user
-    # delegate :published_contracts,  :to => :user
-    # delegate :taken_contracts,      :to => :user
-    # delegate :xid,                  :to => :user
-
-    def initialize(args)
-      xargs = args.except *%i(password)
-      @user = User.find_by(xargs) || User.create(args)
+    def initialize(xargs)
+      args = xargs.stringify_keys
+      args["encrypted_password"] = pwd_digest(args["password"])
+      args["amount"]             = args["balance"] if args["balance"]
+      args["uuid"]               = args["uuid"] || SecureRandom.uuid
+      add_event(:usr1, Event::UserCreated.new(usr_opts(args)))
+      add_event(:usr2, Event::UserDeposited.new(deposit_opts(args))) if has_amount?(args)
     end
 
-    def event_data
-      user.attributes
+    def user
+      @usr2 || @usr1
+    end
+
+    private
+
+    def pwd_digest(password)
+      User.new("password" => password).encrypted_password
+    end
+
+    def usr_opts(opts)
+      cmd_opts.merge(opts.slice(*%w(uuid email encrypted_password)))
+    end
+
+    def deposit_opts(opts)
+      cmd_opts.merge(opts.slice(*%w(uuid amount)))
+    end
+
+    def has_amount?(args)
+      args["amount"]
     end
   end
 end
