@@ -1,6 +1,6 @@
 class Repo::Cvrf < Repo
 
-  # has_many :bugs, :dependent => :destroy
+  # has_many :issues, :dependent => :destroy
 
 end
 
@@ -10,13 +10,12 @@ end
 #
 #  id         :integer          not null, primary key
 #  type       :string
+#  uuid       :string
 #  name       :string
-#  json_url   :string
-#  html_url   :string
+#  xfields    :hstore           not null
 #  jfields    :jsonb            not null
 #  synced_at  :datetime
-#  exref      :string
-#  uuref      :string
+#  exid       :string
 #  created_at :datetime         not null
 #  updated_at :datetime         not null
 #
