@@ -1,31 +1,62 @@
-require 'net/http'
+require 'open-uri'
 
 class Repo::GitHub < Repo
 
-  validates :json_url  , uniqueness: true, presence: true
-  validates :name     , uniqueness: true, presence: true
-
-  before_validation :set_url
-
+  validates :name, uniqueness: true, presence: true
   validates :name, format: { with:    /\A[\_\-\.a-zA-Z0-9]+\/[\.\_\-a-zA-Z0-9]+\z/,
                              message: "needs GitHub repo '<user>/<repo>'" }
 
-  validate :repo_url_presence
+  hstore_accessor  :xfields , :languages  => :string
+  jsonb_accessor   :jfields , :readme_url => :string
+  jsonb_accessor   :jfields , :readme_txt => :string
+
+  validate :repo_presence
+
+  after_create :set_languages
+  after_create :set_readme
+
+  # ----- INSTANCE METHODS
+
+  def html_url
+    "https://github.com/#{self.name}"
+  end
+
+  def readme_type
+    case readme_url.split(".").last
+      when "md"       then "markdown"
+      when "markdown" then "markdown"
+      else "text"
+    end
+  end
+
+  def readme_html
+    self.readme_txt
+  end
+
+  def org
+    self.name.split("/").first
+  end
 
   private
 
-  def set_url
-    self.html_url = "https://github.com/#{self.name}/issues"
-    self.json_url  = "https://api.github.com/repos/#{self.name}/issues"
+  def set_languages
+    return if Rails.env.test?
+    languages = Octokit.languages(self.name)
+    update_attribute :languages, languages.to_hash.keys.map(&:to_s).join(", ")
   end
 
-  def repo_url_presence
-    lcl = set_url
-    url = URI.parse(lcl)
-    req = Net::HTTP.new(url.host, url.port)
-    req.use_ssl = true
-    res = req.request_head(url.path)
-    return if res.code == "200"
+  def set_readme
+    return if Rails.env.test?
+    readme_url = Octokit.readme(self.name)[:download_url]
+    return if readme_url.blank?
+    update_attribute :readme_url, readme_url
+    readme_txt = open(readme_url) { |io| io.read }
+    update_attribute :readme_txt, readme_txt
+  end
+
+  def repo_presence
+    repo = Octokit.repo(self.name)
+    return if repo.present?
     errors.add :name, "GitHub repo does not exist"
   end
 end
@@ -36,13 +67,12 @@ end
 #
 #  id         :integer          not null, primary key
 #  type       :string
+#  uuid       :string
 #  name       :string
-#  json_url   :string
-#  html_url   :string
+#  xfields    :hstore           not null
 #  jfields    :jsonb            not null
 #  synced_at  :datetime
-#  exref      :string
-#  uuref      :string
+#  exid       :string
 #  created_at :datetime         not null
 #  updated_at :datetime         not null
 #
