@@ -17,9 +17,34 @@ require 'rspec/rails'
 # If you are not using ActiveRecord, you can remove this line.
 ActiveRecord::Migration.maintain_test_schema!
 
+require 'vcr'
+
+VCR.configure do |c|
+  c.cassette_library_dir = "spec/_vcr"
+  c.hook_into :webmock
+  c.configure_rspec_metadata!
+end
+
+module AuthRequestHelper
+  BASE_MAIL = "test@bugmark.net"
+  BASE_PASS = "bugmark"
+
+  def create_user(email = BASE_MAIL)
+    FB.create(:user, email: email, password: BASE_PASS)
+  end
+
+  def basic_creds(email = BASE_MAIL)
+    args = [email, BASE_PASS]
+    el = ActionController::HttpAuthentication::Basic.encode_credentials(*args)
+    {'HTTP_AUTHORIZATION' => el}
+  end
+end
+
+USE_VCR = {vcr: {allow_playback_repeats: true}}
+
 RSpec.configure do |config|
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
-  config.fixture_path = "#{::Rails.root}/spec/fixtures"
+  # config.fixture_path = "#{::Rails.root}/spec/fixtures"
 
   # If you're not using ActiveRecord, or you'd prefer not to run each of your
   # examples within a transaction, remove the following line or assign false
@@ -48,5 +73,9 @@ RSpec.configure do |config|
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
 end
+
+
+
+
 
 
